refactor: remove int8/int16/int32/float32

+++ internal/database/offline/postgres/type_tag_test.go
@@ -17,7 +17,7 @@ func TestValueTypeTag(t *testing.T) {
 		},
 		{
 			"int",
-			types.INT32,
+			types.INT64,
 		},
 		{
 			"varchar(32)",
